+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Version history for Skipto.js
 
+## Version 5.6.3
+* Changed default highlight option to be 'instant' instead of 'none'
+* Updated illinois configuration to have highglight option be 'none'
+* Updated highlight div creation to be dynamic
+
 ## Version 5.6.2
 * Changed default highlight option to be 'disabled', added 'smooth' and 'instant' to replace 'enabled' option